requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
PyQt5>=5.15.0
pyserial>=3.5
requests>=2.26.0
python-telegram-bot>=13.7 